requirements.txt
pymongo[srv]==4.5.0
python-telegram-bot==20.5
python-dotenv==1.0.0
waitress